requirements.txt
pandas==1.2.5
python-decouple==3.8
requests==2.22.0
